install_server: change partition scheme for Hadoop workers with GPUs
The new Hadoop workers with GPUs have 24x2TB disks, and no flex bay.
We usually manually configure the extra storage disks (so no need for
them to be present in partman) but in this case, we'd need to reserve
the first two to have a raid1 layout to place the root.

Bug: T254892
Change-Id: I3309e2c4cfcbbff0861c4212f6526958d582e2c1