tools-grid: Install correct version of php-igbinary
It looks like the version of php-igbinary seems to be a bit scattered
on exec nodes because it is installed as a dependency and not intentionally.

On many nodes, we get 2.0.1 (mostly newly made nodes that are missing
some backports) and older onces have 2.0.7, usually. The dev bastion (-08)
has 2.0.8, which is installed from the WMF PHP 7.2 component. I think that
is the correct version to have everywhere, and this should see to that.

Bug: T262186
Change-Id: I1d794f8bec2ff07b8a224e7bd3e634e14847cbd5