role::elasticsearch::cirrus: add missing mjolnir hiera value
Even if mjolnir is not running on the hosts anymore (absented config)
the new hiera value is needed by the profile. We might need to rollback
to the old settings for mjolnir on es nodes, so it is better to keep the old
config updated in case we need it. This also fixes puppet errors currently
ongoing for all the es nodes.

Bug: T260305
Change-Id: I9b695187ee89eec00c392a068327a4fe67c3bfb8