Update prometheus-amd-rocm-stats.py to handle new driver readings
In T260442, we updated the kernel driver for AMD GPUs on stat1005, and
discovered that the new driver makes rocm-smi emit new/different
temperature readings when compared to the old shipped-with-kernel
driver.

This change updates our Prometheus helper script to handle the new
readings correctly, while keeping the old code functional. However, due
to how labels in Prometheus work, the old reading will result in a new
timeseries (all labels on a timeseries must be populated).

Bug: T262404
Change-Id: Id8a6a1051511c2a8edfd6417a35eb830a8e4a08d