arclamp: fix spurious "environment changed"
Swift-enabled cron jobs require multiple environment variables to be
set.  When I added these in Id017fe76, I used a newline-separated
string, as this is how the environment appears in the crontab.

When Puppet checks to see if a crontab has changed, it reads the
environment variables into an *array* of strings.  A newline-separated
string will thus never match, and Puppet will rewrite the crontab
entry on every run.  This is reported in syslog as due to "environment
changed."

This patch switches to using an array to hold the variables, which
fixes the comparison.  Puppet will handle writing each array member on
a separate line for us.

This also adds the Swift variables to the arclamp-compress-logs job,
since the final version of Ifa834869 included code to copy the logs to
Swift in addition to the SVGs.

Bug: T244776
Change-Id: I0d25704a365d31b6b17e3ae20bae01fa6924bef4