mariadb: Add arbcom_ruwiki to the list of private wikis
arbcom_ruwiki is about to be created as a private wiki. As such,
it should not be publicly available on cloud. Add the wiki to
the list of private wikis, so the appropiate filters can be added
to replication on both datacenters.

It will require a restart of sanitarium hosts (at least for s5) on
both dcs.

Bug: T262832
Change-Id: I9fbe64b7111d520f6acabb0bac68ff712b4d9b9b