wikireplicas: Work toward a proxy setup on multi-instance replicas
The general idea here is that physical haproxy servers will perform a
simple, port-based proxy. An additional VM-based proxy (not on this patch)
will access these using 8 VIPs, corresponding to sections.
That will allow the WMCS DNS scripts, which already map database CNAMES
in DNS to section A records like: s1.analytics.db.svc.eqiad.wmflabs.
Each of these, could point to the appropriate vips instead of just the
primary IP of the proxy, allowing listen segments of the config file to
direct to the correct backend DB instances. This removes the need to
rewrite as much code for WMCS users to connect to nonstandard ports.
They just have to connect to already-existing DNS entries. This will
still burden users with some changes, but with
an easier learning curve and less fiddling. I believe this is the original
intent for the wiki replica DNS setup.

The other notion I'm putting forward is that the base config is directed by the
instances defined at the database servers rather than via complex hiera
(if I did that reduce function against a puppetdb query correctly).
This would simplify proxying to 8 servers with differing
section instances. Also, this allows admins to use hiera to override
backends with everything evenly balanced as the default.

There is quite possibly a mistake in the config building here (last time
I used puppetdb queries like this, I was missing a flatten()), but
on merge, it will become apparently pretty quickly. The additional config
file will not cause a reset, so it will be safe to revert and work on.

Bug: T260389
Change-Id: Iee32e82031518b1e9b90e53f6ed36a45464d9c11