mariadb: Increase max_allowed_packet to 64MB on all generic misc dbs
OTRS recommends increasing it to 64MB, and I think nowadays, it is not
a terrible suggestion. Because we cannot change it for OTRS only at the
moment, apply it to all non-special misc hosts (which should be ok).

This should also fix warnings about some row/field size being too large
for the package size.

Bug: T187984
Change-Id: If0f6a1b854cf99f2a104a9e63e2a08898e5e8e25