openstack: replace hiera() with lookup()
This is one very larget patch but it strictly just
replaces hiera() with lookup() where there are no default
values and leaves any other improvements to separate changes.

The strategy is to get rid of a LOT of uses of hiera() at once
without having to worry about it breaking anything since it should
really stay noop and there is little room for error.

This was created with just "sed -i" recursively and then reverting
a small handful of cases that have default values. So almost no
manual editing.

Follow-up patches will remove the remaining special cases and might
add other improvements such as data types but will be much easier
to read and review.

Bug: T209953
Change-Id: Id63cfe55d1a9cb0bb470c1775df34782ea001660