cumin: add alias for just canary appservers
follow-up to I351db1504974

As pointed out by Effie we should have an alias
for just MediaWiki appserver canaries after we extended
the previous mw-canary alias to also include API servers.

Technically "mw" should be "mw-api" and "mw" should be "mw-app + mw-api"
but that is left as it is in this patch.

Change-Id: I66998c47cb3b994a3909d602e336bf4de7bdfe94