cloudelastic: use envoy to mitigate tls latency
Cloudelastic physically speaking lives in eqiad, and we have recently
cut over the active dc to codfw as part of the yearly dc switchover. As
a result, the fact that cloudelastic must perform a new tls handshake
for every request introduces something like 3x codfw RTT of delay into
each request.

Sticking envoy in front of it will take advantage of the connection pool
feature and allow us to restore adequate performance.

Bug: T263073
Change-Id: I2ffba1b23e2fece2c767680b66c1ef5cd49bb265