mailman: require package pwgen to create random passwords
pwgen was installed on the previous mailman list server,
fermium, but apparently it was not puppetized so it's not
installed on the current server lists1001.

SREs doing clinic duty sometimes have to reset list passwords
when admins are not reachable anymore or have lost access.

It's useful to have this tool around to automatically create
random passwords which get mailed to admins.

As mentioned on https://wikitech.wikimedia.org/wiki/Mailman#TLDR
which fails without this being around.

Change-Id: Icab0d890118781102978f8a9a8c1d5937a8e3b4b