Remove the Yubikey HSM classes
This was once installed a prototype with a few users, but there are better
solutions these days (native U2F support in OpenSSH).

Also, the solution is mostly neglected/abandoned by Yubico; it was never
ported to Python 3, the last commit is from January 2019 and it was
eventually removed from Debian by the maintainer (who also works for
Yubico).

As such, remove the Puppet classes from git, they can always be pulled
from git history if we need them again.

Change-Id: I2a3d8e138d1fdeffcf6d58d002454441b60afaf1