mariadb-backups: Start using the package instead of the script on puppet
The check that does freshness testing for mariadb backups at icinga
has been traditionally installed from puppet. Now start using
the better source-controlled wmfbackups-check package, on the
operations/software/wmfbackups repo.

Remove the old logic, including the file from puppet.

This script doesn't depend on wmfmariadbpy, however, it will probably
do it soon, due to its need of section checking, so making the
dependency explicit to the wmfmariabpy library, even if not needed
at the time of this commit. This is beacuse of:
https://gerrit.wikimedia.org/r/c/operations/software/wmfbackups/+/628172

Bug: T138562
Bug: T257551
Change-Id: If93e74757ead869c78b625fd1f3ac30bebeb5091