mariadb-backups: Use the wmfbackups-remote package and not the puppet version
Remove remote_backup_mariadb.py from puppet and install instead wmfbackups-remote
from the debian package.

Also move configuration from /etc/mysql/backups.cnf to
/etc/wmfbackups/remote_backups.cnf

This one doesn't need to be set as a library because it is already an 'admin'
role, so it will install both the library and the admin_cli packages.

Bug: T138562
Bug: T257551
Change-Id: I3c68c880bb7d6d37eff0b627499f1967812543b3