cloudgw: add templated network interfaces file
Instead of using the interface:: modules which may be better for this particular case given the
network configuration is easier to represent just writting the file instead of dealing with augeas.

This also includes the VRF setting to ensure traffic flowing to/from the cloud never hits the
cloud-hosts-vlan and we can use double default routes.

Bug: T261724
Change-Id: I620107485402cafb8f769655903d060a58572f6a
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>