cloudgw: interfaces: more template fixes
So this template includes some changes that are actually meant for augeas to perform (the IPv6 set
in profile::standard). This results in a conflict at puppet agent run time. Previous patches
removed the lines from here but actually the fix might be to leave the template exactly as augeas
would produce it.

The other change is a standard templating thing: extra white spaces and lack of newlines.

Bug: T261724
Change-Id: I0798cb0569d51ed3d270fda1bdcdede5214eab17
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>