cloudgw: hardcode mapped IPv6 interface
The pre-up command that sets the token wont work for a virtual interface. It works with physical
NICs because they are detected by udev and a link is created (in down state though).

You can set the token for a DOWN interface but you can't for a yet-to-be-created interface like
a vlan tagged one in the cloudgw case.

Bug: T261724
Change-Id: I9d0673e1b8af99191a337116382059498465381f
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>