service_proxy: add ipv6 config option on services_proxy config
What?
Add an option to allow services_proxy to listen to both
ipv4 and ipv6

Why?
'localhost' resolves to ::1 by default, but services_proxy
currently is listening to 127.0.0.1, so we get a
connection error, increasing number of
"failed connection attempts". This is not causing any problems
per se, but looking at graphs of this metric, can yield
real connectivity issues within our infra.

Bug: T255568
Change-Id: If4719d88abebc6005478367f2f3d958972da62fc