Add a helper to dump/restore memcached for reboots
We currently replicate all memcached state changes to eqiad/codfw to be able
to failover. If we however reboot an IDP node, the current session state is
lost, so add a helper script to save/restore state for reboots, the session
data is small enough here.

Tested on idp-test by logging in with an active session, dumping state,
restarting memcached and restoring, after which my SSO session continued to
work.

Bug: T233933
Change-Id: I04c4922acfd1adde38d784825fc43b522830b42d