mariadb-backups: Reorganize remote backups (snapshots) for speedup
Based on the average speed of performing a backup, reorganize
the place where these are taken so we optimize for busy resources.

This is the average time (in minutes) that took for the following
backups to execute this month, and the chosen dbprov host:
+---------+----------+
| section | duration |
+---------+----------+
| x1      |  41.0000 | -> 3
| s8      | 283.5714 | -> 1
| s7      | 257.7857 | -> 2
| s6      | 110.1786 | -> 3
| s5      | 115.7241 | -> 2
| s4      | 259.9643 | -> 1
| s3      | 252.5000 | -> 3
| s2      | 165.3929 | -> 3
| s1      | 186.1429 | -> 2
+---------+----------+

Note x1 has to be on dbprov1003 as it is the only section already
upgraded to buster.

This will be tested and this patch may be reverted later this day,
once the backups run correctly.

Bug: T257551
Change-Id: Iefb498c9f0cbb43e8abf05ea7d611e86c6ff3c7c