Drop MemorySwapMax=0 from analytics puppet roles
This setting does not do what it is intended to do (at least on Debian
Buster):

[.../puppet-override.conf:7] Memory limit '0' out of range. Ignoring.

I talked with Luca and we decided that the setting isn't really
necessary. If we find it is, we need to figure out how to disable Swap
some other way. The documentation at
https://freedesktop.org/software/systemd/man/systemd.resource-control.html
is not very helpful.

Change-Id: Icaf3beda1ac77b93fcb3f3bb3d2237799f3d6013