cirrus: temporarily disable saneitizer
The cloudelastic cluster has fallen behind on it's update queue and
needs some help catching up. Pause the saneitizer to reduce the rate
at which indexing requests are being created.

To allow this an `ensure` parameter was added to periodic_job. The
underlying timer implementation already supports ensure so this is
a simple pass-thru.

Bug: T263073
Change-Id: Ia51f1386dfb00ceb4885a19cfd006e64f0ec2db1