Remove obsolete nda_audit script
This came up during the Py2->Py3 porting efforts. But at this point we can
simply remove the script entirely.

This script was used for audits between what was in LDAP and the records
of Legal, but in 2016 we changed the process so that these are now enforced
on an ongoing manner: Since then all users with privileged LDAP are stored
in data.yaml and there's a daily cron which does cross checks against the
content of the various privileged LDAP groups (and more).

Bug: T247364
Change-Id: I3aac3b1e51596e58f25e32cccd4046c2e25a8493