VCL: Attach a variety of GeoIP info as bereq headers; test GeoIP
For requests to intake-logging.wikimedia.org, attach a bunch of
GeoIP data to the backend request as header values.  (From there,
eventgate-wikimedia can parse it out into fields in the NEL event.)

This also involves opening the 'ISP' database, which has AS number
and ISP/organization name data.

Also, we add some MaxMind-provided test databases to the Vagrant
environment, so we can perform basic tests of the new codepaths here.
Later we should add some testing of the geoip cookie feature.

Bug: T263496
Change-Id: I498a0c4af056f979a4a9e7819dca30fbbc2fdc40