mariadb::core_test: convert role to profile, add data types
A role class should not have parameters and hiera calls.

Split into a role and a profile class, move the parameters
to the profile class, keep the includes and role description
in the role class.

Also replace hiera() with lookup() for future puppet6 support.

This is a noop except one detail. The DB cluster name can't be
in the role description this way.

https://puppet-compiler.wmflabs.org/compiler1003/25484/db1133.eqiad.wmnet/index.html

Fixes 8 style issues and introduces one.
wmf-style: total violations delta -7

Change-Id: Id1aadc44d25a9d0b42ddb138fa8a4a7919bb0e94