logstash: keep at least 2 copies of each shard for apifeatureusage
apifeatureusage is managed by logstash, but is hosted on the
production Search cluster. It is meant to serve user traffic
and should be robust. We need to keep at least 2 copies of each
shard to ensure that no data is lost. Also, we don't care about
HDD vs SSD on the Search cluster, only SSDs are available.

To make it more explicit that apifeatureuasage is a different use
case, a new curator configuration file is introduced, with a
configuration adapted for apifeatureusage.

Bug: T263970
Change-Id: I157ff341aec9a719222a1521c833c776897fda6e