interfaces: drop aggregate (bonding) dead code
The only use this code had that I know of, is when we were creating bondings for our cloud
hypervisors. That is no longer our setup.

This code is unused, and has several shortcomings that prevents it from working in modern Debian.
Better to drop dead code instead of having it laying around.

Change-Id: I39bd682f3b55f7ddc546315e9a0336b78e134966
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>