openstack: neutron: use iptables-legacy
Apparently iptables-nft-restore has issues loading a ruleset produced by the linuxbridge-agent in
cloudvirt nodes. The easiest workaround is to use iptables-legacy instead.

Even though cloudnet servers don't present this issue, I'm switching them too for the sake of
consistency.

This patch cleans up a similar workardoung in the nova-compute manifest, and also stop installing
the nft package which is no longer needed if using this workaround.

Bug: T262979
Change-Id: I5697f262fd98b929edec361e3213fb980df8df8b
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>