hieradata: enable onhost memcached on mw2271
After testing on mwdebug1001, we can proceed and 
run on a canary production server. 

This mcrouter configuration actually will have 
no effect until we push the relevant mediawiki 
changes where mediawiki will add a specific key 
prefix to any key we are interested in storing 
in the onhost memcached. 

With this patch though memcached and its exporter
will be installed on mw2271, which is what we need
for a production test. 😋

Bug: T263958
Change-Id: Ife457c651da45a87463e72e655e18b56ccd05e9b