Generalize profile::statistics::gpu
This change adds a new profile called 'profile::amd_gpu' that is meant
to be used on stat100x hosts and hadoop workers (equipped with gpus).
On Debian Stretch the 'render' group is not present (due to a lower version
of systemd) so the kfd device needs a special udev rule to be able to be
used by analytics-privatedata-users. Another alternative could be to deploy
systemd 2.4.1 from the specific component, but the final result should be
the same (every user in analytics-privatedata-users is placed in 'render'
via admin module).

Bug: T255138
Change-Id: Ia0ac7400babdde547bb4cefb4d9e4590c4085648