openstack: cloudgw: introduce native vlan for easier reimaging
We need to leave the cloud-hosts- vlan out of the port trunk for easier reimaging. Otherswise,
the debian installer wont work (missing DHCP) because the switch will recv untagged packets.

With this change, the native vlan for the cloudgw server is the cloud-hosts-xxx vlan, and the
switch will be configured to accept untagged packets for this vlan.

Bug: T263622
Bug: T261724
Change-Id: Ib4232ca2515c1cb2fe94a8d89f9edde86572f61e
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>