dnsdist: add acl parameter for webserver configuration
dnsdist 1.5.0 adds an "acl" parameter to the webserver configuration
that specifies the list of IP addresses that are allowed to open a
connection to it. This defaults to "127.0.0.1, ::1".

Since we use dnsdist's web server to export data to Prometheus, this
commit adds an option to the webserver_config type to accept a list of
IP addresses (CIDR netmasks) that are allowed to access the server,
setting the value to 0.0.0.0/0 and ::/0.  Access to the server is
already restricted by the firewall to traffic from production networks
(see commit 96cbb659a).

Bug: T263789
Change-Id: I3a3e9f3e2b65bfde947e1b955c957c01aa948acf