import IP address type patterns from stdlib 5.2
In stdlib versions before 5.2 the regex used to match a valid
IP in CIDR notation does not allow a /0 as a netmask.

Sukhe ran into this issue when trying to use "0.0.0.0/0" and puppet
compiler would fail to validate it as an IP.

Importing just these 4 files from 5.2 as suggested by Brandon.

Later we can upgrade the entire stdlib to 6.x.

Change-Id: If72501610fee82429395830c66d57a5a4d0972e7