consolidate bastionhost roles, remove module
After I26c74a6a5047bd the difference between a "general"
and a "pop" bastionhost will be gone.

The "opsonly" bastionhost is not in use in production
and I can't see it in openstack-browser but it's possible
it is used in cloud for the "restricted" bastion?

But if it is..there still seems no code difference between them.

Remove the previous roles and just use a simple "bastionhost"
on all of them.

Update cumin alias accordingly.

Also the bastionhost module does nothing except installing
a few packages, so moved that to a profile class. It does not
have to be an entire module.

Change-Id: I67363d7a76decb4caba8018db08eefda94335603