admins/bd808: add bash shebang to .bash scripts
The background to this is the old task T95064 which
made me use the tool "checkbashisms" from the "devscripts"
Debian package and run it across the puppet repo after
a couple years have passed.

One thing that made the output a bit noisy were these 3 scripts
because they did not have a shebang line but it detected
a lot of "possible bashisms" in them.

Adding the #!/bin/bash shebang fixes it and removes a bunch of
noise leaving the remaining cases across the repo.

The full output can be seen at:
https://phabricator.wikimedia.org/P12910

Bug: T95064
Change-Id: I8d58f5138a12efec338b61a83220a6e3c1801f9d