admins/ori: add bash shebang to .z.sh
The background to this is the old task T95064 which
made me use the tool "checkbashisms" from the "devscripts"
Debian package and run it across the puppet repo after
a couple years have passed.

One thing that made the output a bit noisy was this script
because it did not have a shebang line but checkbashisms detected
a lot of "possible bashisms" in them.

Adding the #!/bin/bash shebang fixes it and removes a bunch of
noise leaving the remaining cases across the repo.

I realize this looks like a zsh script but
https://github.com/rupa/z
also says:

AVAILABILITY
       bash, zsh

The full output can be seen at:
https://phabricator.wikimedia.org/P12911

Bug: T95064
Change-Id: I47f857c159beec5ce3ac6a4cd227714c2d535b02