stdlib: update to version 5.2.0
A follow-up after If72501610fee8242 and comments there.

Update stdlib to 5.2.0 because we already took the IP
patterns from it but did not update the other files
and we don't want to mix versions.

The current release is already 6.5.0 but this would be
a first step in that direction.

Last time we updated stdlib was in 2018.

Change-Id: I40f9d9fe73f5c7a851ea74191accd4fc41ed506a