dynamicproxy: guard against missing Host header
Both domainproxy and urlproxy perform lookups based on the Host header
sent by the client. Without a target hostname, neither proxy is able to
determine which upstream server to direct the request to. This change
adds an explicit check for a missing Host header and returns an HTTP 400
"Bad Request" response. This should reduce noise in the nginx logs
resulting from malformed (from the point of view of the service)
requests.

Change-Id: I7b8bf42e4e6905f47e8fef7f500051de4bc1b47b