redis::instance: raise the file limit to match maxclients
Redis allocates a maximum number of connections that is equal to the
minimum between its maxclients setting and the total number of
filehandles it can allocate, minus 32. By default, maxclients is
10000, but the maximum number of open files per systemd service is
4096, resulting in all of our redis instances allowing at most 4064
concurrent connections. That created issues in one specific
case (ORES) but we should more in general honour the redis
configuration and its defaults.

Bug: T263910
Change-Id: I4919234249ce7b28745d6cecd673b62ed2dcb3bf