pontoon: write the stack name once to the filesystem
The idea is to "prime" the stack name within /etc/pontoon-stack. After that the Pontoon server is no
longer dependent on a branch being pushed with 'pontoon::stack' set at all. The main use case /
scenario is protecting against accidental pushing of branches where the stack is not set at all.

Therefore once a Pontoon server has a stack set, a stack change (which is meant to be a rare event)
has to be performed by changing /etc/pontoon-stack directly.

Change-Id: I2573f3c16f74eca41b8bc2752667764922c0e565