diffscan: add defeat-rst-ratelimit
While investigating the long running network scans i noticed that the
scanning the lvs hosts seemed to take a lot of time.  After tweaking a
bit I noticed that adding the `--defeat-rst-ratelimit` flag sped the
scans up significantly.  From `1262.84 seconds` to  `47.71 seconds`.

The setting its self make no difference to us as as it just removes the
ability to distinguish between closed and filtered ports

nmap uses the following states for a SYN port scan:
  * closed - received a RST or icmp port prohibited
  * filtered - no packets recived
  * open - recived an ACK

At a high level if nmap starts scanning starts receiving some RST
packets it (loosely) assumes there is no filtering and therefore we should
receive a RST or an ACK packet from each port probed.

If nmap doesn't see theses packets it assumes they have been dropped on the
network due to either some type of ratelimiting or lossy network and
starts to slow down the rate at which it sends packets with a message
like the following

 Increasing send delay for 192.0.2.42 from 5 to 10 due to 17 out of 42
 dropped probes since last increase.

If a RST packet is never recived this dynamic slow down doesn't happen
and the normal timeout options are honored.  Adding the
`--defeat-rst-ratelimit` switch makes nmap behaviour similar to if it
never received a RST packet

Bug: T264888
Change-Id: Ib12783bfa702d0f15ecf2388ba8426f27657852f