map::postgresql_common: make maps-admin chgrp toggle
Currently puppet is failing in deployment-prep because the maps-admin
group has not been created because we don't apply the Admin class
there. Given that users will have more extensive access in
deployment-prep and LDAP groups are not created in the first place,
removing this chgrp in some cases makes sense.

Bug: T263726
Change-Id: I2bccacb496438a4615609921b0610903b4ea076b