Add an apt proxy config for deb.debian.org
By default we use the Debian mirror we provide to the world for package
installations and updates (sans security.debian.org for security updates
and apt.wikimedia.org for internal packages).

Add deb.debian.org (which is based on the Fastly CDN instead of single
mirror servers) in addition to the apt proxy config. This is needed for

- failover in case sodium (which is a single, non redundant server) gets
  reimaged or is broken in some way (/etc/apt/sources.list is now managed
  via Puppet)
- for the installation of dbgsym packages provided by the debian-debug
  suite

Bug: T262647
Bug: T164819
Change-Id: Ibabb0909ddb87e79f2d4e2717cde735c8c8b04b7