amd_rocm: Ensure linux-headers-amd64 is installed
rock-dkms quietly skips compiling the kernel module if we don't
have the headers for the revelant kernels installed. So before we
add the ROCm packages to the machine, install the kernel headers.
We don't fold this install into the package list since we can't
rely on apt/dpkg getting the ordering right.

Change-Id: If57baae774ab8bd997f2cacd0629fe244fc0009b