VCL: temp. ratelimit iOS app fetches of random page summary
This is a temporary mitigation until the iOS app team has pushed
a new release and that release is widely installed.
This ought to be able to be deleted by 2020-10-23 at latest.

At local midnight, current versions of the iOS app perform several
fetches.  Most of them are cacheable, but the random page summary isn't.
This causes excessive load on several different backend systems, and
often creates pages at 22:00 UTC, local midnight for CEST.

This is a global ratelimit, not a per-IP ratelimit, with a threshold
chosen to allow lots and lots of headroom for usual traffic, but that
will cut off the steepest peaks that we see: https://w.wiki/g3W
We also assume that the traffic hits only one cluster at a given time,
since, well, the traffic is sent at local midnight.

Bug: T264881
Change-Id: I422e86ef2447d2348fdd8857d27c2e4e7d25647d