puppetmaster: pass $servers parameter to gitclone class
The $servers parameter already exists in class puppetmaster
which uses this.

Just pass it along and we don't need to do another Hiera lookup
for the same thing in the module which means the lint ignore
can also be removed.

Additionally the "include puppetmaster" seems redundant
since puppetmaster is the (only) class using puppetmaster::gitclone.

Change-Id: I6b76b2e7351b8b58764797248c5453a830a94097