parsoid: stop using nodejs parsoid on scandium
As described by Moritz on T257906 the goal is to
have the nodejs service on the new buster VM testreduce1001
while removing it from the previous test host scandium on stretch.

The profile::parsoid class included here uses class parsoid
which uses 'service::node' which pulls in node.

We want to reinstall scandium _without_ node but keeping other
parts.

Bug: T257906
Change-Id: I454b495e6d8096405a45c60cd3c05e030d2dac60