varnish: do not install -dbg package
Starting with 6.0.6-1wm2, which solves T264074, we are leaving the
creation of packages with debugging symbols to debhelper instead of
explicitly creating varnish-dbg ourselves:
https://gerrit.wikimedia.org/r/c/operations/debs/varnish4/+/634028

This means that the debug package has now been renamed from varnish-dbg
to varnish-dbgsym. Given that the debugging symbols for Varnish are only
necessary when debugging specific issues and are in general not
necessary on production nodes, they can just be installed when needed.

Stop installing varnish-dbg.

Bug: T264074
Change-Id: I71ec16452fdeb665ac3d025fe2bfdacc3efab92b