allow secteam-users access to nodes with role(peek)
This machine was setup by Chase who had global root access.
There was no additional admin group applied on it though.

Now secteam-users should be able to SSH to it without needing
global root.

Also moving the existing email setting from profile/codfw-level
to role level. That way things are in one place and a matching
VM can be created in eqiad in the future.

Bug: T265922
Change-Id: I4a01ce2d70608abd93a35a313fb0159bdbd542dc