etherpad: add commitRateLimiting config, set higher values
   * From Etherpad 1.9.0 onwards, when Etherpad is in production mode commits from
     individual users are rate limited
   *
   * The default is to allow at most 10 changes per IP in a 1 second window.
   * After that the change is rejected.

Add the new config option.

Set the duration to seconds instead of 1 and the number of changes per IP
to 10 instead of 100.

The expectation is that Etherpad sees all edits as coming from envoy because
that does TLS termination hence causing the rate limiting issues users
have reported since our upgrade to 1.8.6.

Bug: T265490
Change-Id: Ie68c87cbce64ee1ea8241efbfbedeac44b12162d