prometheus: add Pushgateway profile and module
The Pushgateway must be run as a singleton, thus introduce 'pushgateway_host' variable for mutual
exclusion. Prometheus itself will pick up the right host when it changes via the standard Puppet
class based mechanism.

For other hosts to reach the Pushgateway the idea is to use a discovery.wmnet CNAME. Unfortunately
that also means that the Puppet and DNS changes need to be coordinated when the Pushgateway needs to
change hosts.

Bug: T249311
Change-Id: I8dc7283d08361f5eb4b3aba10e9dfa5a6ebb1e4f