dnsdist: set and increase the value of setMaxTCPClientThreads
Since dnsdist 1.4.0, a single TCP worker can handle more than one
connection simultaneously (TCP connection handling was changed to
event-based in 1.4.0), so this is different from previous versions of
dnsdist in which the number of simultaneous TCP connections was limited
to the value set by setMaxTCPClientThreads.

dnsdist sets the maximum number of TCP client threads to 10 by default.
We should allow for setMaxTCPClientThreads to be set in the dnsdist
module and increase it to 20 to start with as recommended by the dnsdist
developers but more experimentation and testing is required to find the
right value for our dnsdist/Wikidough instance(s). (Our current stats
indicate that 20 worker threads are sufficient for the current number of
connections.)

Change-Id: I90b93f7c78badbb1f017bb3d474eb48abf77e56c