openstack: pdns recursor: allow querys from extra CIDRs
Introduce an additional hiera key for cases in which we want to allow extra CIDRs to query the
recursor DNS servers.

This is required if we want to allow CIDRs not present in the central network data yaml, for
example, with CloudVPS floating IP CIDRs.

While at it, drop production networks from the list of accepted recursors clients. It was allowed
in the pdns configuration but missing from the firewalling config anyway.

Bug: T261724
Change-Id: I6d027363075f7dd04e6958f55a3a6c81d5b03ea5
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>