cloud nfs: close all unnecessary ports
Since big changes in the NFS servers around the period of
I8fe9cbb84331c527cf36 , we have been on a strictly NFSv4 system.
Theoretically, none of these other services should require remote access
as a result. In fact, most of these ports are not even listening.
This should be carefully rolled out to cloudstore1009 first.

Bug: T265588
Change-Id: Ie3f8d7614f936c1f23905ee5e7a1c9c750ee2dc8