dumps nfs: remove probably-unused firewall ports and services
This needs to be checked carefully. All cloud clients mount via NFSv4.
That doesn't guarantee *all* clients do. However, NFSv4 does not need
any external access to the removed ports.

This is also setting statd to not start (not needed for v4) for the
dumps distribution servers (labstore1006/7). In fact, those servers are
currently set to not even support v3 (as servers anyway).

Bug: T265588
Change-Id: I101f694b057e9913106279d75806b93182c76248