turnilo/netflow: add more TCP flag combinations
Add more TCP flag values as seen in the wild. Generated using a quick
Python script (P13050), which unfortunately changed the order of fields
from the previous mapping.

We could explore in the future parsing the bitfield using Druid's
Javascript support, but from the docs it looked experimental and
disabled by default, so a map will have to do for now.

Change-Id: I34f91da0f75069523dd68426b17d68ca9bbaca14