profile::hive::client: refactor code to have settings only in one place
The hive settings need to be in sync between server and clients in all roles
that need it. This refactoring is also useful in case Analytics needs to failover
the Hive server node to another host, to avoid replacing the same value across
multiple hiera configs in puppet.

Drop also old code that is not used anymore.

Bug: T257412
Change-Id: Icd619eb64ef2416b6f31d94d88bc79651f74a69f