mediawiki: Check number of cached keys in php-check-and-restart.sh
Why:
max_cached_keys is the max number of keys in opcache. When the
num_cached_keys reaches this max, opcache restarts and clears
all keys. We recently noticed that after this restart, a server
may start erroring.

What:
php-check-and-restart.sh already triggers a php-fpm restart in
case a server is low on free opcache or issue an apcu clear
command if it has a high percentage of APCu fragmentation.
I added one more check that triggers a php-fpm restart if
num_cached_keys is getting near max_cached_keys.

Bug: T253673
Change-Id: Icb00848eb3a19366b84a9730fab8669477e19335