toolforge: script to make long-running processes on bastions less good
This would be a systemd timer on bastions that would look for things run
by login users that have been running for a long time, weight them
based on how long they have been around and randomly selecting some to
be killed. This means a long script could still finish, but it strongly
discourages leaving services around on bastions.

Bug: T266300
Change-Id: If0e22d971d2a15994ff2adb9852be84605cd8f10