safe-service-restart: add optional poolcounter support
Poolcounter is the distributed locking system we use in
production. Let's eliminate a footgun (for now, only optionally) by
adding poolcounter support to the safe restart script.

For now, two conditions need to be met to add poolcounter support to
your safe-service-restart procedure:
- python3-poolcounter is available
- you explicitly set --max-concurrency to a non-zero value

For poolcounter to work, you also need a client configuration to be
available in yaml format.

Bug: T266055
Change-Id: I448d756dbe1da5fc064a165e490700b8921865d1