Set debian buster for mc2036
Why:
It is high time we upgraded a memcached server to buster

What:
Which version of redis will be installed when we move to buster
is still to be determinedi (T265643). Since we will reimage one host
on eqiad and one on codfw, we have decided it is ok if we do
not install redis on those two hosts.

Bug: T252391
Change-Id: Ib30bf128a4a4a1aaf1ba10a31253ad49754f3b78