orchestrator: add monitoring for process and TCP port
Adds a profile for monitoring of the MySQL orchestrator
daemon.

Checks for running process and open TCP port.

Makes host and port configurable with defaults matching the current
setup on dborch1001.

Adds the ability to enable/disable for individual hosts / DCs / roles
via Hiera. Defaults to disabling the checks so one has to opt-in to enable it
but at the same time there are defaults for host and port. (Some might
say no defaults at all would be even better fwiw).

The "notes_url" on Wikitech does not actually exist yet but should be created
and put in Category:Runbooks ideally.

Bug: T266338
Change-Id: I680e1bf01bc093b465e77263a855da83e24396b7