dumps: rm profile::dumps::distribution::datasets::cleanup_miscdatasets
Since this cronjob is hardcoded to "absent", can we just remove
this entire class and stop including it?

If it should be kept because eventually it should be "present" again
then I would like to instead convert it to a systemd timer because
in Puppet 6 cron() will be deprecated. And then we should make
$ensure a parameter.

Bug: T265138
Change-Id: Ifaaf08a415fed378faba5374889e565d3088b662