releases: only use auto-restart script on primary server
We have Icinga alerts on releases2002 about failed systemd units.

One of them is wmf_auto_restart_rsync.service. It fails because the
rsync unit is not present: Active: inactive (dead) which is expected
on the non-primary server.

Move the auto-restart code so that it only applies on the primary
server where the actual rsyncd is installed as well to avoid this.

Change-Id: I636c030a97714028f01af8e5e5661cf892e41e02