remove scandium from scap dsh group
scandium had puppet disabled for a while
which was requested to be able to make some tests.

After a week it fell out of puppetDB which means
during a scap run it will pop up as a warning because
host keys can't be verified.

Also we can now reimage scandium and until then removing
it here to avoid the warnings for deployers.

Bug: T257906
Change-Id: I24ae289112086b9b3455b197a40b8d88e0772172