ores: Stop memory reporting
The error reporting was added in I5daa80256dd to debug the memory issue in T242705 but that issue
has been addressed and resolved now. Also, this reporting is not acually working because of the way
uwsgi formatts the log output, it omits the memory reporting. It reduces profiling overhead in ores.

Change-Id: Ibc9b096bc22d8574096b36a0fd2683c3a9da2958