site/appservers: cleanup comments about appserver rack locations
I would like to remove all the "mw* are in rack" comments.

Servers tend to move and these easily become wrong
and/or it's extra work to keep them updated while the regexes
right below are already containing that information and
of course we have netbox and can always check there.

At the same time, I still would like to keep one section
per rack and not completely drop that part of the information.

To keep it nice and merge-able I am sticking to comments in this
patch. There was a previous attempt but it got lost in rebase hell
and was trying to do too much at once.

So not trying to merge regexes to have exactly one per rack
and sorting them numerically. But that could be done afterwards.

Change-Id: I3e14d5149f1b8a142a16afa73bed4c23740c945c